changed some types and casting. Fixed issue with detecting num_bytes read and written

+++ src/engine/tile/tilemap.h
@@ -19,8 +19,8 @@ struct AutoTiledTileMap
     std::vector<int> tileset_sprite_sheet_indices;
     std::vector<bool> walls; // used to determine auto tiling for the tileset_sprite_sheet_indices
     Uint16 tile_size;
-    Uint16 n_rows;
-    Uint16 n_cols;
+    Uint32 n_rows;
+    Uint32 n_cols;
 };
